Create class `BinanceProcessor`, move all structs to header file. Abstracts the processesing of received Binance Kline JSON objects *away from* the client websocket that receives them. Allows for testing by making the logic more "public" - if you will.
